refactor: useCommunity hook
